Replace local variables with instance variables in tests
We’d like to extract these tests into modules so we can reuse them, but
that’ll take them out of the lexical scope of local variables like `ns`,
`_Object` and `_Class`. By assigning those local variables’ values to
instance variables, and referring consistently to instance variables
throughout, we’re able to remove the requirement for the tests to be
defined lexically within a metaprogramming loop.

To aid in this decoupling, I mostly made the decision to name the
instance variables after the abstract role they play rather than the
concrete value they take: `@class` in the object and module tests is the
class of the instance being tested, and `@superclass` in the class test
is likewise the superclass of the class being tested. I hope this makes
it clearer why these variables exist and how they should be used in
(existing and future) tests.

However, both the object and class tests need access to `Class`
specifically as a factory for generating new anonymous classes, so I’ve
broken that responsibility out into a distinct `@Class` instance
variable in both places for the sake of clarity. This instance variable
is directly populated with `ns::Class` to indicate that it should always
have the same value independently of which object is currently under
test.