Refer directly to `Object` in class test loop
This loop is unusual because it only has one iteration, but regardless,
the subject of that iteration is the `singleton`, just as in the object
and module tests.

The fact that `Object` is `@class`’s superclass is more incidental. It’s
really a consequence of the test setup code, i.e. our decision to
populate `@class` by calling `Class.new`, so to me it feels clearer to
explicitly set `Object` as the superclass during that particular setup
rather than injecting it as one of the parameters of the iteration as if
it has the same significance as `Class`.

(This borderline pathological hair-splitting would be retroactively
justified if we ever wanted to separately test `@class = singleton`,
because then the superclass would decidedly not be `Object` any more —
the `Class` class is an edge case in that regard.)