Use `singleton` as the local variable name in test metaprogramming loops
Now that the tests depend upon instance variables, we’re only referring
to this local variable within the test setup. I think it’s a little
clearer to give it a special name to highlight its role in the test,
i.e. it’s the specific singleton whose behaviour we’re testing, rather
than a more concrete (and now potentially misleading) name which makes
it look like it must necessarily be `Object`, or `Module`, or whatever.

This leaves us in the situation where the remaining capital-letter
variable names (i.e. `_Object` and `@Class`) really do refer to that
class alone (i.e. `Toy::Object` and `Toy::Class`), which I hope is less
confusing.