Singletons should have names
Instead of #inspect in each of the singletons returning a stringified
version of the class, we should give each singleton a name. We can't
rely on ObjectInstance's implementation of #inspect to inspect the singletons,
because this creates a circular chain (calling #inspect calls self.class,
which calls #to_s on the class, which in turn calls inspect). As such, we
define #inspect at the ModuleInstance level, because we know that Module instances
and Class instances can have names and thus need a custom inspect method.
This custom #inspect method opts to use the instance's name, if one exists,
and otherwise defers responsibility to ObjectInstance to #inspect the object
using its class.

Note that we must provide a default implementation of #name on ModuleInstance (nil),
because newly generated Module instances won't have a name (it's only our singletons
that we have defined the #name method for). A client could define name themselves on
a Module or Class instance, but just as in real Ruby, an anonymous Class / Module
should return `nil` when asked for it's name.