Support ObjectInstance#method method
This method returns an object (Toy::ObjectInstance::Method) that responds to #owner.
The method object is "bound" to the object instance, meaning that calling #owner on
it will return the class | module where that is method is defined _for that object instance_.

For now, we achieve this by walking up the superclass hierarchy and asking the class for its
instance methods, checking if one of them is the selector. If we find a method,
we init a Method object whose owner is that class. If we walk the whole superclass hierarchy
and don't find the method, we raise a NameError akin to how Ruby would.